refactor(frontend: rename file
