+++ gui/frontend/src/services/api/electron/index.ts
@@ -1,0 +1,7 @@
+export const isElectron = () => {
+  //@ts-ignore
+  return (globalThis || window).__ELECTRON__ !== undefined;
+};
+
+// @ts-ignore
+export const electronApi = window.__ELECTRON__;
